+++ src/Hyperwallet/Model/PayPalAccount.php
@@ -28,6 +28,10 @@ class PayPalAccount extends BaseModel {
      */
     private static $READ_ONLY_FIELDS = array('token', 'status', 'createdOn');
 
+    public static function FILTERS_ARRAY() {
+        return array('status');
+    }
+
     /**
      * Creates a instance of PayPalAccount
      *